backend/requirements.txt
# Python dependencies for the Syntra FastAPI backend.
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
google-api-python-client==2.127.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
